% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_mzr_header}
\alias{check_mzr_header}
\title{Check that mzr header is as expected}
\usage{
check_mzr_header(x)
}
\arguments{
\item{x}{mzr header data.table}
}
\value{
TRUE/FALSE
}
\description{
Check that mzr header is as expected
}
