% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_mzr_object}
\alias{check_mzr_object}
\title{Check if variable is a mzR obect}
\usage{
check_mzr_object(x)
}
\arguments{
\item{x}{mzR object}
}
\value{
TRUE/FALSE
}
\description{
Check if variable is a mzR obect
}
