% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_if_found.R
\name{color_if_found}
\alias{color_if_found}
\title{Find m/z that match db}
\usage{
color_if_found(
  ms_data,
  npatlas,
  cols,
  ppm,
  select_column = "No Filter",
  select_attr = NULL,
  offset = 1.00784
)
}
\arguments{
\item{ms_data}{mzr connection and header}

\item{npatlas}{npatlas as data.table}

\item{cols}{m/z index}

\item{ppm}{dalton difference, not ppm}

\item{select_column}{optional column filter}

\item{select_attr}{optional filter value(s)}

\item{offset}{adduct mass}
}
\value{
index
}
\description{
Find m/z that match db
}
