% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_header.R
\name{connect_to_ms}
\alias{connect_to_ms}
\title{Connect to an MS file with mzR (Alias for mzR::openMSfile)}
\usage{
connect_to_ms(filename, backend = NULL, verbose = FALSE)
}
\arguments{
\item{filename}{
    Path name of the netCDF, mzData, mzXML or mzML file to
    read/write.
  }

\item{backend}{
    A \code{character(1)} specifiying which backend API to
    use. Currently 'Ramp', 'netCDF'  and 'pwiz' are supported. If
    \code{backend = NULL} (the default), the function tries to determine
    the backend to be used based on either the file extension of the
    file content.
  }

\item{verbose}{ Enable verbose output. }
}
\value{
mzR connection
}
\description{
Connect to an MS file with mzR (Alias for mzR::openMSfile)
}
