% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_header.R
\name{create_header}
\alias{create_header}
\title{Create a header from x-levels of MS connection}
\usage{
create_header(mzr_connection, level = 2L)
}
\arguments{
\item{mzr_connection}{mzR connection}

\item{level}{MS1, MS2 etc. as integer vector}
}
\value{
mzR header as data.table
}
\description{
Create a header from x-levels of MS connection
}
