% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_fileInput.R
\name{mod_fileInput_ui}
\alias{mod_fileInput_ui}
\alias{mod_fileInput_server}
\title{mod_fileInput_ui and mod_fileInput_server}
\usage{
mod_fileInput_ui(id)

mod_fileInput_server(input, output, session, ms_object)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
