% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_gnps_masst_search.R
\name{mod_gnps_masst_search_ui}
\alias{mod_gnps_masst_search_ui}
\alias{mod_gnps_masst_search_server}
\title{mod_gnps_masst_search_ui and mod_gnps_masst_search_server}
\usage{
mod_gnps_masst_search_ui(id)

mod_gnps_masst_search_server(input, output, session, ms_object)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
