% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_mass_rt_plot.R
\name{mod_mass_rt_plot_ui}
\alias{mod_mass_rt_plot_ui}
\alias{mod_mass_rt_plot_server}
\title{mod_mass_rt_plot_ui and mod_mass_rt_plot_server}
\usage{
mod_mass_rt_plot_ui(id)

mod_mass_rt_plot_server(
  input,
  output,
  session,
  ms_object,
  ppm,
  select_column,
  select_attr,
  offset
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
