% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_result_table.R
\name{mod_result_table_ui}
\alias{mod_result_table_ui}
\alias{mod_result_table_server}
\title{mod_result_table_ui and mod_result_table_server}
\usage{
mod_result_table_ui(id)

mod_result_table_server(
  input,
  output,
  session,
  ms_object,
  ppm,
  select_column,
  select_attr,
  offset
)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
