% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mzml_header.R
\name{mzml_header}
\alias{mzml_header}
\title{Create a mzR:header but make data.table}
\usage{
mzml_header(ms_connection)
}
\arguments{
\item{ms_connection}{An instantiated mzR object.}
}
\value{
data.table mzR header
}
\description{
Create a mzR:header but make data.table
}
