% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_npatlas.R
\name{read_npatlas}
\alias{read_npatlas}
\title{Read www.npatlas.org data included within this package}
\usage{
read_npatlas()
}
\arguments{
\item{path}{path to database tsv}
}
\value{
data.table
}
\description{
Read www.npatlas.org data included within this package
}
