% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm.R
\name{scale_ppm}
\alias{scale_ppm}
\title{Currently a passthrough}
\usage{
scale_ppm(mass, ppm)
}
\arguments{
\item{mass}{mass vector}

\item{ppm}{ppm vector}
}
\value{
numeric vector
}
\description{
Currently a passthrough
}
