% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_npatlas_search.R
\name{search_npatlas}
\alias{search_npatlas}
\title{Search a mass against np atlas}
\usage{
search_npatlas(
  npatlas,
  ms_object,
  ppm,
  select_column,
  select_attr,
  offset,
  selected
)
}
\arguments{
\item{npatlas}{npatlas data.table}

\item{ms_object}{mzr connection and header}

\item{ppm}{dalton difference, not currently ppm difference}

\item{select_column}{optional column to filter by}

\item{select_attr}{optional value to filter}

\item{offset}{adduct mass offset}

\item{selected}{selected MS event}
}
\value{
data.table
}
\description{
Search a mass against np atlas
}
